Refactor configuration management and error handling in payment processing
