+++ src/main.ts
@@ -1,7 +1,6 @@
-import { ValidationPipe } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import type { MicroserviceOptions } from '@nestjs/microservices';
-import { RpcException, Transport } from '@nestjs/microservices';
+import { Transport } from '@nestjs/microservices';
 
 import { AppModule } from './app.module';
 
@@ -14,17 +13,6 @@ async function bootstrap() {
         host: '0.0.0.0',
       },
     },
-  );
-
-  app.useGlobalPipes(
-    new ValidationPipe({
-      whitelist: true,
-      forbidNonWhitelisted: true,
-      disableErrorMessages: true,
-      exceptionFactory: errors => {
-        return new RpcException(errors);
-      },
-    }),
   );
   await app.listen();
 }
